requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-appconf==1.0.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-compressor==2.2
django-filter==1.1.0
djangorestframework==3.8.2
idna==2.6
kombu==4.2.1
pytz==2018.4
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
urllib3==1.22
vine==1.1.4
